Updating the type of the argument of RequestHandler
And adding a little bit of refactor in the process
